+++ cider-interaction.el
@@ -796,7 +796,7 @@ The handler simply inserts the result value in BUFFER."
                                '()))
 
 (defun cider-emit-interactive-eval-output (output)
-  "Emit standard or error output resulting from interactive code evaluation.
+  "Emit output resulting from interactive code evaluation.
 
 The output can be send to either a dedicated output buffer or the current REPL buffer.
 This is controlled via `cider-interactive-eval-output-destination'."
@@ -806,6 +806,20 @@ This is controlled via `cider-interactive-eval-output-destination'."
                       (cider-emit-into-popup-buffer output-buffer output)
                       (pop-to-buffer output-buffer)))
     (`repl-buffer (cider-repl-emit-interactive-output output))
+    (t (error "Unsupported value %s for `cider-interactive-eval-output'"))))
+
+;; TODO: Factor out repeated code
+(defun cider-emit-interactive-eval-err-output (output)
+  "Emit error output resulting from interactive code evaluation.
+
+The output can be send to either a dedicated output buffer or the current REPL buffer.
+This is controlled via `cider-interactive-eval-output-destination'."
+  (pcase cider-interactive-eval-output-destination
+    (`output-buffer (let ((output-buffer (or (get-buffer cider-output-buffer)
+                                             (cider-popup-buffer cider-output-buffer t))))
+                      (cider-emit-into-popup-buffer output-buffer output)
+                      (pop-to-buffer output-buffer)))
+    (`repl-buffer (cider-repl-emit-interactive-err-output output))
     (t (error "Unsupported value %s for `cider-interactive-eval-output'"))))
 
 (defun cider-interactive-eval-handler (buffer)
@@ -818,7 +832,7 @@ This is controlled via `cider-interactive-eval-output-destination'."
                                (lambda (_buffer out)
                                  (cider-emit-interactive-eval-output out))
                                (lambda (buffer err)
-                                 (cider-emit-interactive-eval-output err)
+                                 (cider-emit-interactive-eval-err-output err)
                                  (cider-highlight-compilation-errors buffer err)
                                  (cider-jump-to-error-maybe buffer err))
                                '()))